tests/wpt/web-platform-tests/css/css-values/absolute-length-units-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>absolute length units test</title>
<meta name="author" title="Hwanseung Lee" href="mailto:hs1217.lee@samsung.com">
<link rel="help" href="https://drafts.csswg.org/css-values-3/#absolute-lengths"/>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #t1 { width: 96px; height: 2.54cm; }
    #t2 { width: 2.54cm; height: 25.4mm; }
    #t3 { width: 25.4mm; height: 101.6q; }
    #t4 { width: 101.6q; height: 1in; }
    #t5 { width: 1in; height: 6pc; }
    #t6 { width: 6pc; height: 72pt; }
    #t7 { width: 72pt; height: 96px; }
</style>
</head>
<body>
    <div id="t1"> </div>
    <div id="t2"> </div>
    <div id="t3"> </div>
    <div id="t4"> </div>
    <div id="t5"> </div>
    <div id="t6"> </div>
    <div id="t7"> </div>

    <script type="text/javascript">

    var test_items = ["t1", "t2", "t3", "t4", "t5", "t6", "t7"]

    function getPropertyValue(id, property) {
        var elem = document.getElementById(id);
        return window.getComputedStyle(elem, null).getPropertyValue(property);
    }

    test(function(){
        for (var i = 0; i < test_items.length; i++) {
            assert_equals(getPropertyValue(test_items[i], "width"), getPropertyValue(test_items[i], "height"));
        }
    }, "all width and height should be same");
    </script>
</body>
</html>
